+++ fuzz/fuzz_targets/bnum.rs
@@ -19,7 +19,7 @@ macro_rules! fuzzy {
                       assert_eq!(encoded_len, value.encoded_len());
                       assert!(value.encoded_len() <= <$ty>::MAX_ENCODED_LEN);
 
-                      let consumed = crate::consume_varint(&buf).unwrap();
+                      let consumed = crate::consume_varint(&buf);
                       assert_eq!(consumed, encoded_len);
 
                       let (bytes_read, decoded) = <$ty>::decode(&buf).unwrap();