+++ fuzz/fuzz_targets/chrono_tz.rs
@@ -11,7 +11,7 @@ fuzz_target!(|value: Tz| {
     let mut buf = [0; <Tz>::MAX_ENCODED_LEN];
     let encoded_len = value.encode(&mut buf).unwrap();
     assert!(!(encoded_len != value.encoded_len() || (value.encoded_len() > <Tz>::MAX_ENCODED_LEN)));
-    let consumed = consume_varint(&buf).unwrap();
+    let consumed = consume_varint(&buf);
     assert_eq!(consumed, encoded_len);
 
     let (bytes_read, decoded) = <Tz>::decode(&buf).unwrap();
@@ -22,7 +22,7 @@ fuzz_target!(|value: Tz| {
     let encoded = encode_tz(value);
     assert!(!(encoded.len() != encoded_tz_len(value) || (encoded.len() > <Tz>::MAX_ENCODED_LEN)));
 
-    let consumed = consume_varint(&encoded).unwrap();
+    let consumed = consume_varint(&encoded);
     assert_eq!(consumed, encoded.len());
 
     let (bytes_read, decoded) = decode_tz(&encoded).unwrap();