+++ fuzz/fuzz_targets/num_complex.rs
@@ -29,7 +29,7 @@ macro_rules! fuzzy {
                             let encoded = [< encode_complex_ $ty:snake >](value);
                             assert!(!(encoded.len() != [< encoded_complex_ $ty:snake _len >] (value) || !(encoded.len() <= <Complex<$ty>>::MAX_ENCODED_LEN)));
 
-                            let consumed = consume_varint(&encoded).unwrap();
+                            let consumed = consume_varint(&encoded);
                             assert_eq!(consumed, encoded.len());
 
                             let (bytes_read, decoded) = [< decode_complex_ $ty:snake >](&encoded).unwrap();
@@ -40,7 +40,7 @@ macro_rules! fuzzy {
                             let mut buf = [0; <Complex<$ty>>::MAX_ENCODED_LEN];
                             let encoded_len = value.encode(&mut buf).unwrap();
                             assert!(!(encoded_len != value.encoded_len() || !(value.encoded_len() <= <Complex<$ty>>::MAX_ENCODED_LEN)));
-                            let consumed = consume_varint(&buf).unwrap();
+                            let consumed = consume_varint(&buf);
                             assert_eq!(consumed, encoded_len);
 
                             let (bytes_read, decoded) = <Complex<$ty>>::decode(&buf).unwrap();
@@ -63,7 +63,7 @@ macro_rules! fuzzy {
                             let mut buf = [0; <Complex<$ty>>::MAX_ENCODED_LEN];
                             let encoded_len = value.encode(&mut buf).unwrap();
                             assert!(!(encoded_len != value.encoded_len() || !(value.encoded_len() <= <Complex<$ty>>::MAX_ENCODED_LEN)));
-                            let consumed = consume_varint(&buf).unwrap();
+                            let consumed = consume_varint(&buf);
                             assert_eq!(consumed, encoded_len);
 
                             let (bytes_read, decoded) = <Complex<$ty>>::decode(&buf).unwrap();