+++ fuzz/fuzz_targets/primitive_types.rs
@@ -16,7 +16,7 @@ macro_rules! fuzzy {
                             let encoded = [< encode_ $ty:snake _to>](&value, &mut buf).unwrap();
                             assert!(!(encoded != [< encoded_ $ty:snake _len >] (&value) || !(encoded <= <$ty>::MAX_ENCODED_LEN)));
 
-                            let consumed = consume_varint(&buf).unwrap();
+                            let consumed = consume_varint(&buf);
                             assert_eq!(consumed, encoded);
 
                             let (bytes_read, decoded) = [< decode_ $ty:snake >](&buf).unwrap();
@@ -27,7 +27,7 @@ macro_rules! fuzzy {
                             let mut buf = [0; <$ty>::MAX_ENCODED_LEN];
                             let encoded_len = value.encode(&mut buf).unwrap();
                             assert!(!(encoded_len != value.encoded_len() || !(value.encoded_len() <= <$ty>::MAX_ENCODED_LEN)));
-                            let consumed = consume_varint(&buf).unwrap();
+                            let consumed = consume_varint(&buf);
                             assert_eq!(consumed, encoded_len);
 
                             let (bytes_read, decoded) = <$ty>::decode(&buf).unwrap();