+++ src/primitive_types/v0_13.rs
@@ -115,7 +115,7 @@ mod tests {
               return false;
             }
 
-            let Ok(consumed) = $crate::consume_varint(&buf) else {
+            let Some(consumed) = $crate::consume_varint_checked(&buf) else {
               return false;
             };
             if consumed != encoded_len {
@@ -138,7 +138,7 @@ mod tests {
               return false;
             }
 
-            let Ok(consumed) = $crate::consume_varint(&buf) else {
+            let Some(consumed) = $crate::consume_varint_checked(&buf) else {
               return false;
             };
             if consumed != encoded {