+++ src/ruint_impl/v1.rs
@@ -177,7 +177,7 @@ mod tests_ruint_1 {
               return false;
             }
 
-            let Ok(consumed) = crate::consume_varint(&buf) else {
+            let Some(consumed) = crate::consume_varint_checked(&buf) else {
               return false;
             };
             if consumed != encoded_len {