+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # RELEASED
 
+## 0.10.0 (Aug 12nd, 2025)
+
+- Make `consume_varint` panic
+- Add `consume_varint_checked` and `try_consume_varint` for non-paniking version `consume_varint`
+
 ## 0.9.0 (Aug 9th, 2025)
 
 - Add `ConstEncodeError` and `ConstDecodeError`